requirements.txt
# QR Code File Encoder - Зависимости

# Основные библиотеки для генерации QR-кодов и работы с изображениями
qrcode>=7.4.0
Pillow>=10.0.0
reportlab>=4.0.0
svgwrite>=1.4.3

# Дополнительно для обработки текста
python-dateutil>=2.8.0

# Для декодирования QR-кодов
pyzbar>=0.1.9

# Для архивации (встроены в Python, но можно обновить)
# gzip, bz2, lzma, zipfile - входят в стандартную библиотеку